requirement.txt
mysql-connector
requests
flask
gunicorn